+++ Content/LeagueSandbox-Scripts/Buffs/Ascension/AscBuffIcon.cs
@@ -20,70 +20,12 @@ namespace Buffs
         };
         public IStatsModifier StatsModifier { get; private set; } = new StatsModifier();
 
-        IParticle p1;
-        IParticle p2;
-        IBuff thisBuff;
         public void OnActivate(IAttackableUnit unit, IBuff buff, ISpell ownerSpell)
         {
-            thisBuff = buff;
-
-            PlaySound("Stop_sfx_ZhonyasRingShield_OnBuffActivate", unit);
-            PlaySound("Play_sfx_Leona_LeonaSolarFlare_hit", unit);
-            PlaySound("Play_sfx_Lux_LuxIlluminationPassive_hit", unit);
-            SetBuffToolTipVar(buff, 3, 15.0f);
-
-            p1 = AddParticle(unit, unit, "Global_Asc_Avatar", unit.Position, -1);
-            p2 = AddParticle(unit, unit, "Global_Asc_Aura", unit.Position, -1);
-            AddParticleTarget(unit, unit, "AscForceBubble", unit, size: 3.3f);
-            AddParticleTarget(unit, unit, "CassPetrifyMiss_tar", unit, size: 3.0f);
-            AddParticleTarget(unit, unit, "Rebirth_cas", unit);
-            AddParticleTarget(unit, unit, "TurnBack", unit);
-            AddParticleTarget(unit, unit, "LeonaPassive_tar", unit, size: 2.5f);
-
-            ApiEventManager.OnDeath.AddListener(unit, unit, OnDeath, true);
-
-            //TODO: Add 100% mana/energy cost reduction and 50% Health cost reduction
-            StatsModifier.HealthPoints.FlatBonus = 50.0f * unit.Stats.Level;
-            StatsModifier.AttackDamage.FlatBonus = 12.0f * unit.Stats.Level;
-            StatsModifier.AbilityPower.FlatBonus = 12.0f * unit.Stats.Level;
-            StatsModifier.ArmorPenetration.PercentBonus = 0.15f;
-            StatsModifier.MagicPenetration.PercentBonus = 0.15f;
-            StatsModifier.CooldownReduction.FlatBonus = 0.25f;
-            StatsModifier.Size.FlatBonus = 0.5f;
-            unit.AddStatModifier(StatsModifier);
-            unit.Stats.CurrentHealth = unit.Stats.HealthPoints.Total;
-            //TODO: Add Buff ToolTips updates here once we figure out why buffs don't show up in the Ascension gamemode
-
-            AddUnitPerceptionBubble(unit, 0f, 25000f, TeamId.TEAM_BLUE, true, unit);
-            AddUnitPerceptionBubble(unit, 0f, 25000f, TeamId.TEAM_PURPLE, true, unit);
-
-            //Note: The ascension applies a "MoveAway" knockback debuff on all enemies around.
-            //The duration varies based on the distance (yet unknown) you were to whoever ascended. And the PathSpeedOverride and ParabolicGravity varies based on the duration.
-            //PathSpeedOverride and ParabolicGravity with 0.5 duration: Speed - 1200 / ParabolicGravity - 10.0
-            //PathSpeedOverride and ParabolicGravity with 0.75 duration: Speed - 1600 / ParabolicGravity - 7.0
-        }
-
-        public void OnDeath(IDeathData deathData)
-        {
-            if (deathData.Unit is IMonster xerath)
-            {
-                AddBuff("AscBuffTransfer", 5.7f, 1, null, deathData.Killer, xerath);
-            }
-            if(deathData.Unit is IChampion)
-            {
-                NotifyAscendant();
-                NotifyWorldEvent(EventID.OnStartGameMessage3, 8);
-                NotifyWorldEvent(EventID.OnClearAscended);
-            }
-
-            thisBuff.DeactivateBuff();
         }
 
         public void OnDeactivate(IAttackableUnit unit, IBuff buff, ISpell ownerSpell)
         {
-            RemoveParticle(p1);
-            RemoveParticle(p2);
-            unit.PauseAnimation(false);
         }
 
         public void OnUpdate(float diff)
